requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
pydantic-settings
pytest
pytest-asyncio
google-generativeai>=0.3.1
